Workout Planner
Some functionality completed. swapped to a modal for exercise adding due to severe lag when loading the 2833 exercises into a drop down. now the entire list loads into the modal. only functionality needed is the Create exercise tab. WorkoutPlannerPage itself loads exercises from the db into the page so whenever a new exercise is selected it loads and dispalys. will need a way to delete and move but otherwise good to go.